tests/wpt/web-platform-tests/html/syntax/parsing/quotes-in-meta.html

<!doctype html>
<meta http-equiv="Content-Type" content='charset="windows-1251'>
<meta charset=windows-1250>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<link rel=help href="https://html.spec.whatwg.org/#algorithm-for-extracting-a-character-encoding-from-a-meta-element">
<script>
test(function() {
    assert_equals(document.characterSet, "windows-1250");
});
</script>
